Re-organized project for pypi compatibility
