--- __init__.py
@@ -1,2 +1,0 @@
-from core import *
-from core import SCHEMA_PATH, SCHEMATA, __all__
