+++ ebmlite/__init__.py
@@ -1,0 +1,4 @@
+from core import *
+from core import SCHEMA_PATH, SCHEMATA, __all__
+
+name = "ebmlite"
